MNT: Moving data_store to its own module.
