+++ pydm/data_plugins/pva_plugin.py
@@ -1,7 +1,7 @@
 import os
 PVA_LIB = os.getenv("PYDM_PVA_LIB", "").upper()
 if PVA_LIB == "P4P":
-    from pydm.data_plugins.epics_plugins.pyepics_plugin_component import PyEPICSPlugin
+    from pydm.data_plugins.pva_plugins.p4p_plugin_component import P4PPlugin
     PVAPlugin = P4PPlugin
 elif PVA_LIB == "PVAPY":
     PVAPlugin = None
